pyproject.toml
[tool.poetry]
name = "sport-telegram-bot"
version = "0.1.0"
description = ""
authors = ["habo3hnk <melinbodb@ya.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.6.0"
sqlalchemy = "^2.0.30"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
pydantic = "^2.7.1"
flake8 = "^7.0.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
